fix: profileDialog presence box
