+++ src/renderer/src/components/Modules/NethVoice/BaseModule/Navbar.tsx
@@ -42,35 +42,37 @@ export function Navbar({ onClickAccount }: NavbarProps): JSX.Element {
   if (!account) return <></>
 
   return (
-    <div className="flex flex-row items-center justify-between gap-4 px-4 pt-2">
-      <SearchBox />
-      <div className="flex flex-row min-w-30 gap-2 items-center">
-        <PresenceBadge
-          mainPresence={account?.data?.mainPresence}
-          className={classNames()}
-        />
+    <>
+      <div className="flex flex-row items-center justify-between gap-4 px-4 pt-2">
+        <SearchBox />
+        <div className="flex flex-row min-w-30 gap-2 items-center">
+          <PresenceBadge
+            mainPresence={account?.data?.mainPresence}
+            className={classNames()}
+          />
 
-        <div className={'max-h-8'}>
-          <div
-            className="cursor-pointer dark:focus:outline-none dark:focus:ring-2 focus:outline-none focus:ring-2 dark:ring-offset-1 ring-offset-1 dark:ring-offset-slate-900 ring-offset-slate-50 focus:ring-primaryRing dark:focus:ring-primaryRingDark rounded-full  "
-            onClick={() => {
-              setIsProfileDialogOpen((p) => !p)
-              debouncer('reload_me', onClickAccount, 1000)
-            }}
-          >
-            <Avatar
-              size="small"
-              status={status}
-              src={operators?.avatars?.[account.username] || undefined}
-              placeholderType="operator"
-            />
+          <div className={'max-h-8'}>
+            <div
+              className="cursor-pointer dark:focus:outline-none dark:focus:ring-2 focus:outline-none focus:ring-2 dark:ring-offset-1 ring-offset-1 dark:ring-offset-slate-900 ring-offset-slate-50 focus:ring-primaryRing dark:focus:ring-primaryRingDark rounded-full  "
+              onClick={() => {
+                setIsProfileDialogOpen((p) => !p)
+                debouncer('reload_me', onClickAccount, 1000)
+              }}
+            >
+              <Avatar
+                size="small"
+                status={status}
+                src={operators?.avatars?.[account.username] || undefined}
+                placeholderType="operator"
+              />
+            </div>
           </div>
         </div>
-      </div>
+      </div >
       <ProfileDialog
         isOpen={isProfileDialogOpen}
         onClose={() => setIsProfileDialogOpen(false)}
       />
-    </div >
+    </>
   )
 }
